backend/pyproject.toml
[tool.poetry]
name = "autoappsec-backend"
version = "0.1.0"
description = "Backend for AutoAppSec"
authors = ["Admin <admin@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = "^2.0.25"
asyncpg = "^0.29.0"
pydantic-settings = "^2.1.0"
alembic = "^1.13.1"
celery = "^5.3.6"
redis = "^5.0.1"
python-multipart = "^0.0.9"
pyjwt = "^2.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
httpx = "^0.26.0"
docker = "^7.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
typer = "^0.9.0"
jinja2 = "^3.1.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.0"
isort = "^5.13.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
